Rename `migrateCalculatedMeasuresInMdx` and refactor
